+++ packages/entities/entities-plugins/src/components/free-form/Datakit/flow-editor/node-forms/NodeFormJq.vue
@@ -1,34 +1,77 @@
 <template>
   <Form
+    ref="form"
+    :config="{ updateOnChange: true }"
+    :data="formData"
     :schema="JqNodeSchema"
-    @change="console.log"
   >
-    <Field name="name" />
+    <StringField
+      name="name"
+      @update:model-value="setName"
+    />
 
     <KLabel class="dk-node-configuration-label">
       {{ i18n.t('plugins.free-form.datakit.flow_editor.node_properties.Configuration') }}
     </KLabel>
 
+    <!-- todo(zehao): replace to monaco editor -->
     <StringField
       :label="i18n.t('plugins.free-form.datakit.flow_editor.node_properties.jq.label')"
       multiline
       name="jq"
       resizable
       rows="2"
+      @update:model-value="setConfig()"
     />
-    <InputsField :items="[]" />
 
-    <!-- Todo(zehao): outputs definition fields -->
+    <InputsField
+      :items="inputOptions"
+      @add:field="handleAddField"
+      @change:input="setInput"
+      @change:inputs="setInputs"
+      @remove:field="handleRemoveField"
+      @rename:field="handleRenameField"
+    />
+
+    <!-- todo(zehao): outputs definition fields -->
   </Form>
 </template>
 
 <script setup lang="ts">
 import Form from '../../../shared/Form.vue'
-import Field from '../../../shared/Field.vue'
 import InputsField from './InputsField.vue'
 import { JqNodeSchema } from '../node/mock'
 import useI18n from '../../../../../composables/useI18n'
 import StringField from '../../../shared/StringField.vue'
+import { useTemplateRef } from 'vue'
+import { useNodeForm } from './composables'
+import type { FieldName, IdConnection } from '../../types'
 
 const { i18n } = useI18n()
+
+const formRef = useTemplateRef('form')
+
+const {
+  formData,
+  setName,
+  setConfig,
+  inputOptions,
+  setInputs,
+  setInput,
+  addField,
+  removeFieldByName,
+  renameFieldByName,
+} = useNodeForm(() => formRef.value!.getInnerData())
+
+function handleAddField(name: FieldName, value?: IdConnection | null) {
+  addField('input', name, value)
+}
+
+function handleRemoveField(name: FieldName) {
+  removeFieldByName('input', name)
+}
+
+function handleRenameField(oldName: FieldName, newName: FieldName) {
+  renameFieldByName('input', oldName, newName)
+}
 </script>