+++ packages/entities/entities-plugins/src/components/free-form/Datakit/flow-editor/node-forms/composables.ts
@@ -1,7 +1,7 @@
-import { computed, nextTick, ref, watch } from 'vue'
+import { computed, nextTick, watch } from 'vue'
 import { useEditorStore } from '../../composables'
 import { buildAdjacency, hasCycle } from '../store/validation'
-import type { EdgeInstance, FieldId, FieldName, IdConnection, NameConnection, NodeId, NodeName } from '../../types'
+import type { EdgeInstance, FieldName, IdConnection, NameConnection, NodeId, NodeName } from '../../types'
 import { findFieldById, findFieldByName, getNodeMeta, parseIdConnection } from '../store/helpers'
 
 export type InputOption = {
@@ -12,7 +12,7 @@ export type InputOption = {
 export type BaseFormData = {
   name: NodeName
   input?: IdConnection
-  inputs?: Record<FieldName, IdConnection>
+  inputs?: Record<FieldName, IdConnection | undefined | null>
 }
 
 export function useNodeForm<T extends BaseFormData = BaseFormData>(
@@ -24,21 +24,21 @@ export function useNodeForm<T extends BaseFormData = BaseFormData>(
     state,
     renameNode,
     getNodeById,
-    addField: addFieldRaw,
-    renameField: renameFieldRaw,
-    removeField: removeFieldRaw,
+    addField: storeAddField,
+    renameField: storeRenameField,
+    removeField: storeRemoveField,
     replaceConfig,
     disconnectEdge,
     connectEdge,
   } = useEditorStore()
 
   const selectedNodeId = computed(() => selectedNode.value!.id)
-  const isGlobalStateUpdating = ref(false)
+  let isGlobalStateUpdating = false
 
   watch(state, async () => {
-    isGlobalStateUpdating.value = true
+    isGlobalStateUpdating = true
     await nextTick()
-    isGlobalStateUpdating.value = false
+    isGlobalStateUpdating = false
   }, { immediate: true, deep: true })
 
   // todo(zehao): debugging store api, remove later
@@ -62,6 +62,14 @@ export function useNodeForm<T extends BaseFormData = BaseFormData>(
       return acc
     }, {})
 
+    // append user defined fields to inputs
+    selectedNode.value?.fields.input.forEach(field => {
+      if (!inputsAndInput.inputs) inputsAndInput.inputs = {}
+      if (!inputsAndInput.inputs[field.name]) {
+        inputsAndInput.inputs[field.name] = null
+      }
+    })
+
     return {
       ...selectedNode.value!.config,
       ...inputsAndInput,
@@ -70,29 +78,55 @@ export function useNodeForm<T extends BaseFormData = BaseFormData>(
   })
 
   const setName = (name: string | null) => {
-    if (isGlobalStateUpdating.value) return
+    if (isGlobalStateUpdating) return
     renameNode(selectedNodeId.value, name as NodeName ?? '')
   }
 
   const setConfig = (commitNow = true) => {
-    if (isGlobalStateUpdating.value) return
+    if (isGlobalStateUpdating) return
     const { name, input, inputs, ...config } = getFormInnerData() as T
     replaceConfig(selectedNodeId.value, config, commitNow)
   }
 
+  const findFieldByNameOrThrow = (
+    io: 'input' | 'output',
+    name: FieldName,
+  ) => {
+    const field = findFieldByName(selectedNode.value!, io, name)
+    if (!field) {
+      throw new Error(`Field with name "${name}" not found in node "${selectedNode.value!.name}"`)
+    }
+    return field
+  }
+
   const addField = (
     io: 'input' | 'output',
-    name: string,
+    name: FieldName,
+    value?: IdConnection | null,
   ) => {
-    addFieldRaw(selectedNodeId.value, io, name as FieldName)
+    const hasValue = !!value
+    storeAddField(selectedNodeId.value, io, name, !hasValue)
+    if (hasValue) {
+      setInputs(name, value)
+    }
   }
 
-  const renameField = (fieldId: FieldId, newName: string) => {
-    renameFieldRaw(selectedNodeId.value, fieldId, newName as FieldName)
+  const renameFieldByName = (
+    io: 'input' | 'output',
+    oldFieldName: FieldName,
+    newFieldName: FieldName,
+  ) => {
+    if (isGlobalStateUpdating) return
+    const fieldId = findFieldByNameOrThrow(io, oldFieldName).id
+    storeRenameField(selectedNodeId.value, fieldId, newFieldName)
   }
 
-  const removeField = (fieldId: FieldId) => {
-    removeFieldRaw(selectedNodeId.value, fieldId)
+  const removeFieldByName = (
+    io: 'input' | 'output',
+    fieldName: FieldName,
+  ) => {
+    const fieldId = findFieldByNameOrThrow(io, fieldName).id
+    storeRemoveField(selectedNodeId.value, fieldId)
   }
 
   const inputEdge = computed<EdgeInstance | undefined>(() => {
@@ -110,10 +144,10 @@ export function useNodeForm<T extends BaseFormData = BaseFormData>(
    *   and add new edges for the input fields.
    */
   const setInputs = (fieldName: FieldName, fieldValue: IdConnection | null) => {
-    if (isGlobalStateUpdating.value) return
+    if (isGlobalStateUpdating) return
     const clearing = fieldValue == null
 
-    const fieldId = findFieldByName(selectedNode.value!, 'input', fieldName)?.id
+    const fieldId = findFieldByNameOrThrow('input', fieldName).id
 
     if (clearing) {
       // remove the edge of the input field
@@ -141,7 +175,7 @@ export function useNodeForm<T extends BaseFormData = BaseFormData>(
       fieldId: sourceField,
     } = parseIdConnection(fieldValue)
 
-    const targetFieldId = findFieldByName(selectedNode.value!, 'input', fieldName)!.id
+    const targetFieldId = findFieldByNameOrThrow('input', fieldName).id
     connectEdge({
       source,
       sourceField,
@@ -156,7 +190,7 @@ export function useNodeForm<T extends BaseFormData = BaseFormData>(
    * - If user sets a new input, disconnect existing edges and add new edge for the input.
    */
   const setInput = (value: IdConnection | null, commitNow = true) => {
-    if (isGlobalStateUpdating.value) return
+    if (isGlobalStateUpdating) return
     const clearing = value == null
 
     // remove existing edges
@@ -227,8 +261,8 @@ export function useNodeForm<T extends BaseFormData = BaseFormData>(
 
     // field ops
     addField,
-    renameField,
-    removeField,
+    renameFieldByName,
+    removeFieldByName,
 
     // input(s) ops
     setInputs,